requirements.txt
requests>=2.28.0
PyJWT>=2.6.0
cryptography>=41.0.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
python-telegram-bot[job-queue]
aiohttp>=3.8.0
aiofiles>=0.8.0
folium>=0.14.0  # 用于创建交互式地图
branca>=0.6.0   # folium的依赖，用于处理颜色渲染
Pillow>=10.0.0  # 用于图像处理
tabulate>=0.9.0  # 用于表格显示